backend/pyproject.toml
[project]
name = "docchat-backend"
version = "1.0.0"
description = "Production RAG API with multi-provider LLM support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.22",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.8.0",
    "python-dotenv>=1.2.1",
    "google-genai>=1.64.0",
    "openai>=1.64.0",
    "groq>=0.14.0",
    "pinecone>=8.1.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-jose[cryptography]>=3.5.0",
    "httpx>=0.28.1",
    "langsmith>=0.7.5",
    "langchain>=1.2.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
